+++ conkit/plot/contactmapchord.py
@@ -41,7 +41,7 @@ __version__ = "0.1"
 import matplotlib.pyplot as plt
 import numpy as np
 
-from conkit.core.contact import ContactState
+from conkit.core.contact import ContactMatchState
 from conkit.plot._figure import Figure
 from conkit.plot._plottools import ColorDefinitions, points_on_circle
 
@@ -149,11 +149,11 @@ class ContactMapChordFigure(Figure):
             # 0.0 transparent through 1.0 opaque
             alpha = float(c[4]) if self.use_conf else 1.0
             color = {
-                ContactState.mismatch: ColorDefinitions.MISMATCH,
-                ContactState.match: ColorDefinitions.MATCH,
+                ContactMatchState.mismatched: ColorDefinitions.MISMATCH,
+                ContactMatchState.matched: ColorDefinitions.MATCH,
             }.get(int(c[5]), ColorDefinitions.MATCH)
             ax.plot(x, y, color=color, alpha=alpha, linestyle="-", zorder=0)
-            if int(c[5]) == ContactState.match:
+            if int(c[5]) == ContactMatchState.matched:
                 ax.plot(x, y, color=color, alpha=alpha, linestyle="-", zorder=1, linewidth=1)
             else:
                 ax.plot(x, y, color=color, alpha=alpha, linestyle="-", zorder=0, linewidth=1)